change ToC config
Set to 10 instead of 250 the minimum number of words to automatically create a ToC. Now the ToC is created and displayed in all the notes that don't have "enableToc: false" in their front matter. Moreover, for a ToC to be displayed there need to be a structure inside the document, i.e. header(s) with "##" and smaller (pages with  just the biggest header  will display an empty ToC.