major changes on file location + obsidian templates
Moved the root of the Obsidian vault INSIDE the /content folder. Images are now linkable with the full path relative to /content and they all must be stored insider /content/images.

Also added and set templates (w/ templater) for session and meeting notes in Obsidian, with callouts formatted in Obsidian core syntax. SET OBSIDIAN TO AUTOMATICALLY CONVERT LINKS INTO MARKDOWN LINKS, instead of wikilinks, so to have a cleaner redering on the website (only the file name, in lieu of the abslute path that was displayed with wikilinks).