naming coventions in templates; timeline; readings
added working timeline to index; set all templates (for sessions, meetings and - via "Citations" settings - for readings notes) to be compatible with my workflow, the creation of the timeline added to index and to have links to Zotero for readings notes.
Also moved to private folder the initial pages made for try.
Naming conventions:
- for readings: automatically set the file name to be equal to the citekey of the document in the Zotero library document
- for sessions: e.g. "session 001" for file name and "Session 001" for title, to be changed everytime upon creation of a new session note.
- for meetings: e.g. "meeting 001" for file name and "Meeting 001" for title, to be changed everytime upon creation of a new meeting note.

With this naming convention notes for sessions and meetings will be easily sortable in the future.
Also, that in the template the big header ( #) has been removed, so that the online page only displays the title, which would be displayed anyways.