fixed image file names
Image files  named with a "@" at the beginning of the file name were not found and consequently not rendered. I adopted a new naming convention:
Before:
"@heibiCrowdsourcingOpenCitations2019-fig3"
Now:
"heibiCrowdsourcingOpenCitations2019-fig3"

For some reason, there is NO PROBLEM with links to "readings" notes, which are also named after the citekey of the paper they're about, therefore include a "@" at the beginning of the file name (e.g. @heibiCrowdsourcingOpenCitations2019)